1. Async load for file duration 2. int to int64 variations fix for x86 systems

+++ videoprocessorprogresspoller.h
@@ -18,7 +18,7 @@ public:
 
 private:
     QTimer *timer;
-    int duration;
+    qint64 duration;
     const int tick = 200;
 
 signals:
